news_scraper_backend/requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
pymongo==4.6.1
python-dotenv==1.0.0
APScheduler==3.10.4
gunicorn==21.2.0
flask==3.0.2
